requirements.txt
discord.py
icalendar
python-dotenv
table2ascii